// src/app/api/api/dividend-metrics.service.ts
export interface ApiFindByYieldApiDividendMetricsSearchGetRequestParams {
    /** Minimum TTM dividend yield (%) */
    minYield?: number;
    /** Maximum TTM dividend yield (%) */
    maxYield?: number | null;
    /** Exact number of dividend payments detected in the rolling TTM window (e.g., 12&#x3D;monthly, 4&#x3D;quarterly) */
    freq?: number | null;
    page?: number;
    pageSize?: number;
}